refactor: move dayMonthYear functions into utils/date
