+++ src/explore-education-statistics-admin/src/pages/release/create-release/CreateReleasePage.tsx
@@ -19,10 +19,8 @@ import {
   errorCodeToFieldError,
 } from '@common/components/form/util/serverValidationHandler';
 import RelatedInformation from '@common/components/RelatedInformation';
-import {
-  emptyDayMonthYear,
-  Publication,
-} from '@common/services/publicationService';
+import { Publication } from '@common/services/publicationService';
+import { emptyDayMonthYear } from '@common/utils/date/dayMonthYear';
 import Yup from '@common/validation/validation/yup';
 import React, { useEffect, useState } from 'react';
 import { RouteComponentProps, withRouter } from 'react-router';
