+++ src/explore-education-statistics-admin/src/pages/release/summary/ReleaseSummaryForm.tsx
@@ -18,8 +18,8 @@ import FormFieldRadioGroup from '@common/components/form/FormFieldRadioGroup';
 import FormFieldSelect from '@common/components/form/FormFieldSelect';
 import FormFieldTextInput from '@common/components/form/FormFieldTextInput';
 import { SelectOption } from '@common/components/form/FormSelect';
-import { DayMonthYearInputs } from '@common/services/publicationService';
 import { Dictionary } from '@common/types';
+import { DayMonthYearInputs } from '@common/utils/date/dayMonthYear';
 import Yup from '@common/validation/validation/yup';
 import { endOfDay, format, isValid } from 'date-fns';
 import { FormikActions, FormikProps } from 'formik';
