+++ src/explore-education-statistics-admin/src/services/dashboard/types.ts
@@ -4,10 +4,8 @@ import {
   IdTitlePair,
   UserDetails,
 } from '@admin/services/common/types';
-import {
-  DayMonthYearValues,
-  ReleaseStatus,
-} from '@common/services/publicationService';
+import { ReleaseStatus } from '@common/services/publicationService';
+import { DayMonthYearValues } from '@common/utils/date/dayMonthYear';
 
 export interface ThemeAndTopics {
   title: string;
